camera_yolo/requirements.txt
ultralytics==8.1.0
opencv-python==4.9.0.80
requests==2.31.0
python-dotenv==1.0.1
numpy